requirements.txt
flask==2.2.2

# Flask extensions
flask-sqlalchemy==3.0.2
flask-marshmallow==0.14.0
flask-jwt-extended==4.4.4
flask-migrate==4.0.2
flask-cors==3.0.10
marshmallow-sqlalchemy==0.28.1

# Third-party libs
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
gunicorn==20.1.0
